BACKLOG-12225 : improvements
