+++ src/javascript/JContent/AppLayout/JContentLogo.jsx
